+++ lib/Besra/Parser/Lit.hs
@@ -2,7 +2,7 @@
 module Besra.Parser.Lit ( parser ) where
 
 import Protolude
-import Besra.Types.IR1.Lit
+import Besra.Types.IR1 ( Lit(..) )
 import Besra.Parser.Helpers ( Parser )
 import qualified Besra.Parser.Char as Char
 import qualified Besra.Parser.String as String
