+++ lib/Besra/Parser/String.hs
@@ -2,7 +2,7 @@
 module Besra.Parser.String ( parser ) where
 
 import Protolude
-import Besra.Types.IR1.String
+import Besra.Types.IR1 ( String(..) )
 import Besra.Parser.Helpers
