+++ lib/Besra/SA/ConflictingBindingDecls.hs
@@ -4,28 +4,27 @@ module Besra.SA.ConflictingBindingDecls ( validate ) where
 import Protolude
 import Besra.SA.Helpers
 import Besra.SA.Types
-import Besra.Types.IR1.Module
-import Besra.Types.IR1.Expr
+import qualified Besra.Types.IR1 as IR1
 import Besra.Types.Id
 import Besra.Types.Ann
 
 
-type Module' = Module Parsed
-type Decl' = Decl Parsed
+type Module' = IR1.Module Parsed
+type Decl' = IR1.Decl Parsed
 
 validate :: FilePath -> Validation [SAError] Module'
-validate path (Module decls) =
+validate path (IR1.Module decls) =
   let typeDecls = filter isBindingDecl decls
       groupedDecls = groupBy sameVar typeDecls
       result = mconcat $ map (checkConflict path) groupedDecls
    in result
 
 isBindingDecl :: Decl' -> Bool
-isBindingDecl (BindingDecl _) = True
+isBindingDecl (IR1.BindingDecl _) = True
 isBindingDecl _ = False
 
 sameVar :: Decl' -> Decl' -> Bool
-sameVar (BindingDecl (Binding _ (Id a) _)) (BindingDecl (Binding _ (Id b) _)) = a == b
+sameVar (IR1.BindingDecl (IR1.Binding _ (Id a) _)) (IR1.BindingDecl (IR1.Binding _ (Id b) _)) = a == b
 sameVar _ _ = False
 
 checkConflict :: FilePath -> [Decl'] -> ValidationResult [SAError]
