+++ lib/Besra/SA/MissingTopLevelDecls.hs
@@ -6,17 +6,15 @@ import Besra.SA.Helpers
 import Besra.SA.Types
 import Besra.Types.Id
 import Besra.Types.Ann
-import Besra.Types.IR1.Module
-import Besra.Types.IR1.Expr ( Binding(..) )
-import Besra.Types.IR1.TypeAnn
+import qualified Besra.Types.IR1 as IR1
 import qualified Data.List as List
 
 
-type Module' = Module Parsed
-type Decl' = Decl Parsed
+type Module' = IR1.Module Parsed
+type Decl' = IR1.Decl Parsed
 
 validate :: FilePath -> Validation [SAError] Module'
-validate path (Module decls) =
+validate path (IR1.Module decls) =
   let matchingDecls = filter isBindingOrTypeAnnDecl decls
       groupedDecls = groupBy sameVar matchingDecls
       result = mconcat $ map (checkConflict path) groupedDecls
@@ -26,18 +24,18 @@ isBindingOrTypeAnnDecl :: Decl' -> Bool
 isBindingOrTypeAnnDecl decl = isBindingDecl decl || isTypeAnnDecl decl
 
 isBindingDecl :: Decl' -> Bool
-isBindingDecl (BindingDecl _) = True
+isBindingDecl (IR1.BindingDecl _) = True
 isBindingDecl _ = False
 
 isTypeAnnDecl :: Decl' -> Bool
-isTypeAnnDecl (TypeAnnDecl _) = True
+isTypeAnnDecl (IR1.TypeAnnDecl _) = True
 isTypeAnnDecl _ = False
 
 sameVar :: Decl' -> Decl' -> Bool
-sameVar (TypeAnnDecl (TypeAnn _ (Id a) _)) (TypeAnnDecl (TypeAnn _ (Id b) _)) = a == b
-sameVar (BindingDecl (Binding _ (Id a) _)) (TypeAnnDecl (TypeAnn _ (Id b) _)) = a == b
-sameVar (TypeAnnDecl (TypeAnn _ (Id a) _)) (BindingDecl (Binding _ (Id b) _)) = a == b
-sameVar (BindingDecl (Binding _ (Id a) _)) (BindingDecl (Binding _ (Id b) _)) = a == b
+sameVar (IR1.TypeAnnDecl (IR1.TypeAnn _ (Id a) _)) (IR1.TypeAnnDecl (IR1.TypeAnn _ (Id b) _)) = a == b
+sameVar (IR1.BindingDecl (IR1.Binding _ (Id a) _)) (IR1.TypeAnnDecl (IR1.TypeAnn _ (Id b) _)) = a == b
+sameVar (IR1.TypeAnnDecl (IR1.TypeAnn _ (Id a) _)) (IR1.BindingDecl (IR1.Binding _ (Id b) _)) = a == b
+sameVar (IR1.BindingDecl (IR1.Binding _ (Id a) _)) (IR1.BindingDecl (IR1.Binding _ (Id b) _)) = a == b
 sameVar _ _ = False
 
 checkConflict :: FilePath -> [Decl'] -> ValidationResult [SAError]
