--- lib/Besra/Types/IR1/Impl.hs
@@ -1,20 +1,0 @@
-
-{-# LANGUAGE UndecidableInstances #-}
-
-module Besra.Types.IR1.Impl ( Impl(..) ) where
-
-import Protolude
-import Besra.Types.IR1.Pred
-import Besra.Types.IR1.Expr ( Binding )
-import Besra.Types.Ann
-import Besra.Types.Span
-
-
-data Impl (ph :: Phase)
-  = Impl (Ann ph) [Pred ph] (Pred ph) [Binding ph]
-
-deriving instance AnnHas Eq ph => Eq (Impl ph)
-deriving instance AnnHas Show ph => Show (Impl ph)
-
-instance HasSpan (Ann ph) => HasSpan (Impl ph) where
-  span (Impl ann _ _ _) = span ann
