--- lib/Besra/Types/IR1/String.hs
@@ -1,10 +1,0 @@
-
-module Besra.Types.IR1.String ( String(..) ) where
-
-import Protolude
-
-
--- | The language only supports text based strings,
---   they are not automatically lists of characters.
-newtype String = String Text
-  deriving (Eq, Show)
