+++ tests/Test/Besra/SA/ConflictingBindingDeclsSpec.hs
@@ -10,11 +10,8 @@ import Besra.SA.Types
 import Besra.Types.Id
 import Besra.Types.Ann
 import Besra.Types.Span
-import Besra.Types.IR1.Module
-import Besra.Types.IR1.Expr
-import Besra.Types.IR1.String
-import Besra.Types.IR1.Number
-import Besra.Types.IR1.Lit
+import Besra.Types.IR1 ( Module(..), Decl(..), Expr(..), Binding(..)
+                       , String(..), Number(..), Lit(..) )
 import Besra.Parser
 import Test.Hspec
